Cast to int32

+++ src/utilities.py
@@ -16,8 +16,8 @@ def projection(gray_img, axis:str='horizontal'):
     """ Compute the horizontal or the vertical projection of a gray image """
 
     if axis == 'horizontal':
-        projection_bins = np.sum(gray_img, 1)    
+        projection_bins = np.sum(gray_img, 1).astype('int32')
     elif axis == 'vertical':
-        projection_bins = np.sum(gray_img, 0)    
+        projection_bins = np.sum(gray_img, 0).astype('int32')
 
     return projection_bins
